# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ClinicalFlow API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginCredentials"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated

  # ── USERS ─────────────────────────────────────────────────────────────────
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: Query attendance records
      parameters:
        - name: scheduleId
          in: query
          schema:
            type: string
        - name: studentId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [present, late, absent, pending]
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
  /attendance/time-in:
    post:
      operationId: recordTimeIn
      tags: [attendance]
      summary: Record GPS + Face Verification time-in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceTimeIn"
      responses:
        "200":
          description: Time-in recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
  /attendance/time-out:
    post:
      operationId: recordTimeOut
      tags: [attendance]
      summary: Record GPS + Face Verification time-out
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceTimeOut"
      responses:
        "200":
          description: Time-out recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
  /attendance/ci-assisted:
    post:
      operationId: ciAssistedAttendance
      tags: [attendance]
      summary: CI records attendance in Clinical Instructor Attendance Mode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CIAssistedAttendance"
      responses:
        "200":
          description: Attendance recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
  /attendance/{id}/manual:
    patch:
      operationId: manualAttendanceOverride
      tags: [attendance]
      summary: CI manually overrides attendance status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceManualUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"

  # ── AVAILABLE DUTY SLOTS ──────────────────────────────────────────────────
...
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List announcements
      responses:
        "200":
          description: Announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create announcement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}:
    patch:
      operationId: updateAnnouncement
      tags: [announcements]
      summary: Edit announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Delete announcement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
  /announcements/{id}/read:
    post:
      operationId: markAnnouncementRead
      tags: [announcements]
      summary: Mark announcement as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Marked read

  # ── ANALYTICS ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    LoginCredentials:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
    AuthUser:
      type: object
      required: [id, email, role, firstName, lastName, isActive]
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, scheduler, ci, student]
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        studentProfile:
          $ref: "#/components/schemas/StudentProfile"
        ciProfile:
          $ref: "#/components/schemas/CIProfile"

    # ── USERS ──────────────────────────────────────────────────────────────
    User:
      type: object
      required: [id, email, role, firstName, lastName, isActive, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, scheduler, ci, student]
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        studentProfile:
          $ref: "#/components/schemas/StudentProfile"
        ciProfile:
          $ref: "#/components/schemas/CIProfile"
...
    StudentProfile:
      type: object
      required: [id, userId, studentNumber]
      properties:
        id:
          type: string
        userId:
          type: string
        studentNumber:
          type: string
        yearLevel:
          type: ["integer", "null"]
        section:
          type: ["string", "null"]
        program:
          type: ["string", "null"]
        academicYear:
          type: ["string", "null"]
        totalHoursRequired:
          type: integer
        createdAt:
          type: string
          format: date-time
    CIProfile:
      type: object
      required: [id, userId]
      properties:
        id:
          type: string
        userId:
          type: string
        employeeId:
          type: ["string", "null"]
        specialization:
          type: ["string", "null"]

    # ── STUDENT DETAILS ────────────────────────────────────────────────────
...
    Hospital:
      type: object
      required: [id, name, isActive, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: ["string", "null"]
        contactNumber:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        attendanceRadius:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        departments:
          type: array
          items:
            $ref: "#/components/schemas/Department"
...
    Department:
      type: object
      required: [id, hospitalId, name, isActive]
      properties:
        id:
          type: string
        hospitalId:
          type: string
        name:
          type: string
        code:
          type: ["string", "null"]
        isActive:
          type: boolean
...
    Schedule:
      type: object
      required: [id, dutyDate, startTime, endTime, status, gracePeriodMin, createdAt]
      properties:
        id:
          type: string
        title:
          type: ["string", "null"]
        hospitalId:
          type: string
        departmentId:
          type: string
        ciId:
          type: string
        dutyDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        gracePeriodMin:
          type: integer
        status:
          type: string
          enum: [upcoming, active, completed, cancelled]
        notes:
          type: ["string", "null"]
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        hospital:
          $ref: "#/components/schemas/Hospital"
        department:
          $ref: "#/components/schemas/Department"
        ci:
          $ref: "#/components/schemas/User"
        students:
          type: array
          items:
            $ref: "#/components/schemas/User"
...
    AttendanceRecord:
      type: object
      required: [id, scheduleId, studentId, status, createdAt]
      properties:
        id:
          type: string
        scheduleId:
          type: string
        studentId:
          type: string
        ciId:
          type: ["string", "null"]
        timeIn:
          type: ["string", "null"]
        timeOut:
          type: ["string", "null"]
        dutyHours:
          type: ["number", "null"]
        status:
          type: string
          enum: [present, late, absent, pending]
        method:
          type: string
          enum: [biometric, ci_assisted, manual]
        studentLatitude:
          type: ["number", "null"]
        studentLongitude:
          type: ["number", "null"]
        gpsVerified:
          type: boolean
        faceVerified:
          type: boolean
        livenessVerified:
          type: boolean
        remarks:
          type: ["string", "null"]
        needsMakeup:
          type: boolean
        makeupCompleted:
          type: boolean
        createdAt:
          type: string
          format: date-time
        student:
          $ref: "#/components/schemas/User"
        schedule:
          $ref: "#/components/schemas/Schedule"
    AttendanceTimeIn:
      type: object
      required: [scheduleId, gpsVerified, faceVerified, livenessVerified]
      properties:
        scheduleId:
          type: string
        studentLatitude:
          type: number
        studentLongitude:
          type: number
        gpsVerified:
          type: boolean
        faceVerified:
          type: boolean
        livenessVerified:
          type: boolean
        deviceInfo:
          type: string
    AttendanceTimeOut:
      type: object
      required: [scheduleId, gpsVerified, faceVerified, livenessVerified]
      properties:
        scheduleId:
          type: string
        studentLatitude:
          type: number
        studentLongitude:
          type: number
        gpsVerified:
          type: boolean
        faceVerified:
          type: boolean
        livenessVerified:
          type: boolean
        deviceInfo:
          type: string
    CIAssistedAttendance:
      type: object
      required: [scheduleId, studentId, status]
      properties:
        scheduleId:
          type: string
        studentId:
          type: string
        status:
          type: string
          enum: [present, late, absent]
        gpsVerified:
          type: boolean
        faceVerified:
          type: boolean
        remarks:
          type: string
    AttendanceManualUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [present, late, absent]
        remarks:
          type: string

    # ── DUTY SLOTS ─────────────────────────────────────────────────────────
...
    Announcement:
      type: object
      required: [id, title, body, targetRole, isPinned, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        body:
          type: string
        targetRole:
          type: string
          enum: [all, student, ci, scheduler, admin]
        postedBy:
          type: ["string", "null"]
        isPinned:
          type: boolean
        expiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        isRead:
          type: boolean
        readCount:
          type: integer
        poster:
          $ref: "#/components/schemas/User"
    AnnouncementInput:
      type: object
      required: [title, body, targetRole]
      properties:
        title:
          type: string
        body:
          type: string
        targetRole:
          type: string
          enum: [all, student, ci, scheduler, admin]
        isPinned:
          type: boolean
        expiresAt:
          type: string

    # ── ANALYTICS ──────────────────────────────────────────────────────────
...